refactor; move favicon-cache to libtransmission
