+++ gtk/FaviconCache.cc
@@ -3,7 +3,7 @@
 // or any future license endorsed by Mnemosyne LLC.
 // License text can be found in the licenses/ folder.
 
-#include "FaviconCache.h"
+#include <libtransmission/favicon-cache.h>
 
 #include <gdkmm/pixbuf.h>
 #include <giomm/memoryinputstream.h>
