+++ libtransmission/favicon-cache.h
